Delete show-detail component and create new show detail component
The previous show-detail component was deleted and replaced by a new show detail component. This new component has better form handling with required validators for status, title, and description. Additionally, it now includes a subscription to return an array of Timezones, Languages, and Categories. This update results in a more efficient and cleaner structure.