Revamp Loading Bar Component and introduction of NgRx Store
This commit includes a major revision of the loading bar component. We have switched to a new version of the loading bar component that uses the NgRx store for dispatching actions. In addition, we have added new GitHub workflow files and modified show creation and detail components along with a range of minor UI tweaks.