Add feature to publish episodes
Imported AlertService in show-episodes component as part of this revision. Also, a new function handlePublishEpisode has been introduced to manage the process of publishing episodes. This function, which is now linked to the "Publish" button in the front-end, uses the updateEpisodeStatus method in the episodes service to switch the status of an episode to 'enabled'. The alertService provides a success message upon successful completion.